% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wingui.R
\docType{data}
\name{GUI}
\alias{GUI}
\title{Windows Rgui accessor class}
\format{An instance of WindowsGUI reference class.}
\usage{
GUI
}
\description{
This object is defined if using the Rgui interface on windows.
Available attributes are available through attributes of \code{GUI}
\enumerate{
   \item \code{$Title} The title of the window.
   \item \code{$opacity} Percentage of the opacity of the window.
   \item \code{$transparency} Percentage of the transparency of the window, wrapper of opacity
   \item \code{$on.top} bollean of if the window is fixed on top.
   \item \code{$layered} bollean of if the window is considered layered.
   \item \code{$.pid} The process ID.
   \item \code{$.hwnd} The window handle.
}
}
\details{
This is a singleton class with instance \code{GUI}.
The is a \link[methods:ReferenceClasses]{reference class} building off the
windows API.
}
\examples{
\dontrun{
GUI$Title
GUI$Title <- "My Title"
GUI$opacity <- 0.90      # set to 90\%
GUI$transparency         # should now be 0.10
GUI$on.top <- T          # Rgui will now always be on top
}
}
\keyword{datasets}

