% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/npptor.R
\name{npp}
\alias{npp}
\title{Launch Notepad++}
\usage{
npp(file = NULL, new = FALSE, exe = getOption("wingui::Notepad++",
  find_npp()))
}
\arguments{
\item{file}{file to open in Notepad++}

\item{new}{open in a new instance?}

\item{exe}{Path to Notepad++ executable.}
}
\description{
Launch Notepad++
}

