% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/npptor.R
\name{npptor}
\alias{npptor}
\title{Launch NppToR}
\usage{
npptor(..., exe = getOption("wingui::npptor", find_npptor()),
  startup = getOption("wingui::startup", is_r_startup()))
}
\arguments{
\item{...}{passed on as arguments to npptor}

\item{exe}{path to the NppToR exacutable}

\item{startup}{should the '-startup' parameter be passed to npptor?}
}
\description{
Launch NppToR
}

