% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/win_read_link.R
\name{read_lnk}
\alias{read_lnk}
\title{Read a windows '.lnk' file.}
\usage{
read_lnk(filename, ...)
}
\arguments{
\item{filename}{file to read.}

\item{...}{discarded}
}
\value{
file path to where the file points, most often specified relatively, along with
		attribute data which contains the information used to parse the lnk file.
}
\description{
Read a windows '.lnk' file.
}

